+++ pkg/compactor/bucket_compactor_e2e_test.go
@@ -42,12 +42,17 @@ import (
 	"github.com/thanos-io/thanos/pkg/block"
 	"github.com/thanos-io/thanos/pkg/block/metadata"
 	"github.com/thanos-io/thanos/pkg/objstore"
+
+	"github.com/grafana/mimir/pkg/storage/tsdb/bucketindex"
 )
 
 const fetcherConcurrency = 32
 
 func TestSyncer_GarbageCollect_e2e(t *testing.T) {
 	foreachStore(t, func(t *testing.T, bkt objstore.Bucket) {
+		// Use bucket with global markers to make sure that our custom filters work correctly.
+		bkt = bucketindex.BucketWithGlobalMarkers(bkt)
+
 		ctx, cancel := context.WithTimeout(context.Background(), 120*time.Second)
 		defer cancel()
 
@@ -112,7 +117,7 @@ func TestSyncer_GarbageCollect_e2e(t *testing.T) {
 
 		blocksMarkedForDeletion := promauto.With(nil).NewCounter(prometheus.CounterOpts{})
 		garbageCollectedBlocks := promauto.With(nil).NewCounter(prometheus.CounterOpts{})
-		ignoreDeletionMarkFilter := block.NewIgnoreDeletionMarkFilter(nil, nil, 48*time.Hour, fetcherConcurrency)
+		ignoreDeletionMarkFilter := NewIgnoreDeletionMarkFilter(nil)
 		sy, err := NewMetaSyncer(nil, nil, bkt, metaFetcher, duplicateBlocksFilter, ignoreDeletionMarkFilter, blocksMarkedForDeletion, garbageCollectedBlocks, 1)
 		require.NoError(t, err)
 
@@ -122,7 +127,10 @@ func TestSyncer_GarbageCollect_e2e(t *testing.T) {
 
 		var rem []ulid.ULID
 		err = bkt.Iter(ctx, "", func(n string) error {
-			id := ulid.MustParse(n[:len(n)-1])
+			id, ok := block.IsBlockDir(n)
+			if !ok {
+				return nil
+			}
 			deletionMarkFile := path.Join(id.String(), metadata.DeletionMarkFilename)
 
 			exists, err := bkt.Exists(ctx, deletionMarkFile)
@@ -161,6 +169,9 @@ func TestSyncer_GarbageCollect_e2e(t *testing.T) {
 
 func TestGroupCompactE2E(t *testing.T) {
 	foreachStore(t, func(t *testing.T, bkt objstore.Bucket) {
+		// Use bucket with global markers to make sure that our custom filters work correctly.
+		bkt = bucketindex.BucketWithGlobalMarkers(bkt)
+
 		ctx, cancel := context.WithTimeout(context.Background(), 120*time.Second)
 		defer cancel()
 
@@ -204,9 +215,9 @@ func TestGroupCompactE2E(t *testing.T) {
 
 		reg := prometheus.NewRegistry()
 
-		ignoreDeletionMarkFilter := block.NewIgnoreDeletionMarkFilter(logger, objstore.WithNoopInstr(bkt), 48*time.Hour, fetcherConcurrency)
+		ignoreDeletionMarkFilter := NewIgnoreDeletionMarkFilter(objstore.WithNoopInstr(bkt))
 		duplicateBlocksFilter := NewShardAwareDeduplicateFilter()
-		noCompactMarkerFilter := NewNoCompactionMarkFilter(logger, objstore.WithNoopInstr(bkt), 2, false)
+		noCompactMarkerFilter := NewNoCompactionMarkFilter(objstore.WithNoopInstr(bkt), false)
 		metaFetcher, err := block.NewMetaFetcher(nil, 32, objstore.WithNoopInstr(bkt), "", nil, []block.MetadataFilter{
 			ignoreDeletionMarkFilter,
 			duplicateBlocksFilter,
@@ -466,6 +477,9 @@ func TestGarbageCollectDoesntCreateEmptyBlocksWithDeletionMarksOnly(t *testing.T
 	logger := log.NewLogfmtLogger(os.Stderr)
 
 	foreachStore(t, func(t *testing.T, bkt objstore.Bucket) {
+		// Use bucket with global markers to make sure that our custom filters work correctly.
+		bkt = bucketindex.BucketWithGlobalMarkers(bkt)
+
 		ctx, cancel := context.WithTimeout(context.Background(), 120*time.Second)
 		defer cancel()
 
@@ -502,7 +516,7 @@ func TestGarbageCollectDoesntCreateEmptyBlocksWithDeletionMarksOnly(t *testing.T
 
 		blocksMarkedForDeletion := promauto.With(nil).NewCounter(prometheus.CounterOpts{})
 		garbageCollectedBlocks := promauto.With(nil).NewCounter(prometheus.CounterOpts{})
-		ignoreDeletionMarkFilter := block.NewIgnoreDeletionMarkFilter(nil, objstore.WithNoopInstr(bkt), 48*time.Hour, fetcherConcurrency)
+		ignoreDeletionMarkFilter := NewIgnoreDeletionMarkFilter(objstore.WithNoopInstr(bkt))
 
 		duplicateBlocksFilter := NewShardAwareDeduplicateFilter()
 		metaFetcher, err := block.NewMetaFetcher(nil, 32, objstore.WithNoopInstr(bkt), "", nil, []block.MetadataFilter{
@@ -546,7 +560,10 @@ func TestGarbageCollectDoesntCreateEmptyBlocksWithDeletionMarksOnly(t *testing.T
 func listBlocksMarkedForDeletion(ctx context.Context, bkt objstore.Bucket) ([]ulid.ULID, error) {
 	var rem []ulid.ULID
 	err := bkt.Iter(ctx, "", func(n string) error {
-		id := ulid.MustParse(n[:len(n)-1])
+		id, ok := block.IsBlockDir(n)
+		if !ok {
+			return nil
+		}
 		deletionMarkFile := path.Join(id.String(), metadata.DeletionMarkFilename)
 
 		exists, err := bkt.Exists(ctx, deletionMarkFile)